+++ src/VisualStudioCode/package/src/configurationFiles.generated.ts
@@ -903,6 +903,9 @@ roslynator_analyzers.enabled_by_default = true|false
 #dotnet_diagnostic.rcs1260.severity = none
 # Options: roslynator_trailing_comma_style
 
+# Dispose resource asynchronously
+#dotnet_diagnostic.rcs1261.severity = suggestion
+
 # Use pattern matching
 #dotnet_diagnostic.rcs9001.severity = silent